backend/FlaskPythonServer/requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
numpy==1.26.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
SQLAlchemy==2.0.29
threadpoolctl==3.4.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
